Cast magma_int_t values to (int) in printf, to avoid warnings when 64-bit. Delete some unused variables. (in magmablas)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3079 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ magmablas/zgetmatrix_transpose_mgpu.cu
@@ -39,7 +39,7 @@ magmablas_zgetmatrix_transpose_mgpu(
 
     if (lda < m || ngpus*ldda < n || lddb < m){
         printf( "Wrong arguments in magmablas_zgetmatrix_transpose_mgpu (%d<%d), (%d*%d<%d), or (%d<%d).\n",
-                lda, m, ngpus, ldda, n, lddb, m );
+                (int) lda, (int) m, (int) ngpus, (int) ldda, (int) n, (int) lddb, (int) m );
         return;
     }
